Split vds.h into types and api
Previously vds.h contained two kinds of code:
- one that was used both by go and internal c++ as a means to translate
types from go to c++ and back using c structs/enums/etc.
- and one that was purely the library "api" - functions called by go and
never by c++ code not directly involved in api implementation.

That made it confusing as vds.h had two types of responsibilities and
internal code needed to know only about types.

Hence the file was split into capi and ctypes files, thus separating the
logic.
