+++ internal/vds/capi.h
@@ -1,6 +1,7 @@
-#ifndef VDS_SLICE_CGO_VDS_H
-#define VDS_SLICE_CGO_VDS_H
-#include <stdlib.h>
+#ifndef VDS_SLICE_CAPI_H
+#define VDS_SLICE_CAPI_H
+
+#include "ctypes.h"
 
 #ifdef __cplusplus
 extern "C" {
@@ -37,12 +38,6 @@ int context_free(Context* ctx);
 
 /** Read out the last error msg set on the context */
 const char* errmsg(Context* ctx);
-
-struct response {
-    char*         data;
-    unsigned long size;
-};
-typedef struct response response;
 
 void response_delete(struct response*);
 
@@ -85,48 +80,6 @@ int regular_surface_free(
     Context* ctx,
     RegularSurface* surface
 );
-
-enum axis_name {
-    I         = 0,
-    J         = 1,
-    K         = 2,
-    INLINE    = 3,
-    CROSSLINE = 4,
-    DEPTH     = 5,
-    TIME      = 6,
-    SAMPLE    = 7,
-};
-
-enum coordinate_system {
-    INDEX      = 0,
-    ANNOTATION = 1,
-    CDP        = 2,
-};
-
-enum interpolation_method {
-    NEAREST,
-    LINEAR,
-    CUBIC,
-    ANGULAR,
-    TRIANGULAR
-};
-
-enum attribute {
-    VALUE,
-    MIN,
-    MAX,
-    MAXABS,
-    MEAN,
-    MEANABS,
-    MEANPOS,
-    MEANNEG,
-    MEDIAN,
-    RMS,
-    VAR,
-    SD,
-    SUMPOS,
-    SUMNEG
-};
 
 int metadata(
     Context* ctx,
@@ -222,4 +175,4 @@ int attribute(
 #ifdef __cplusplus
 }
 #endif
-#endif // VDS_SLICE_CGO_VDS_H
+#endif // VDS_SLICE_CAPI_H
