+++ internal/vds/cppapi_data.cpp
@@ -1,4 +1,4 @@
-#include "vds.h"
+#include "ctypes.h"
 
 #include <cstdint>
 #include <string>
