+++ internal/vds/cppapi_metadata.cpp
@@ -1,4 +1,4 @@
-#include "vds.h"
+#include "ctypes.h"
 
 #include "nlohmann/json.hpp"
