+++ internal/vds/subvolume.cpp
@@ -4,7 +4,7 @@
 
 #include "axis.hpp"
 #include "metadatahandle.hpp"
-#include "vds.h"
+#include "ctypes.h"
 
 namespace {
